// lib/types.ts
export interface DecodedVin {
  battery_type: string;
  body_type: string;
  model: string;
  motor: string;
  vin_locations: string[];
  location_of_manufacture: string;
  restraint_systems: string;
  serial_number: number;
  manufacturer: string;
  year: number;
  vin: string;
  make: string;
  variant: string;
  drive: string;
  trim: string;
  car_images: string[];
}